pyproject.toml
[project]
name = "stampli-disney-reviews"
version = "0.1.0"
description = "Disneyland Reviews RAG API with FastAPI and Gradio UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.119.1",
    "gradio>=5.49.1",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "numpy>=1.24.0",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.0.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.0.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
# Testing dependencies - only installed in development/CI
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.24.0",
]

# Development dependencies - linting, formatting, etc.
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]
